Crate sui_protocol_config_macros

Derive Macros

ProtocolConfigAccessors

This proc macro generates getters, attribute lookup, etc for protocol config fields of type Option<T> and for the feature flags Example for a field: new_constant: Option<u64>, and for feature flags feature: bool, we derive

ProtocolConfigFeatureFlagsGetters

ProtocolConfigOverride